% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_cache_dir}
\alias{load_cache_dir}
\title{load a knitr cache that is NOT lazy}
\usage{
load_cache_dir(path, cache_env, ext = "RData")
}
\arguments{
\item{path}{path to cache dir}

\item{cache_env}{an env that has already been created to capture the objects}

\item{ext}{cache file extension}
}
\description{
qwraps2::lazyload_cache_dir() works if you have set lazy = TRUE in knitr
}
\details{
For very large files this can crash, so I may need to turn off lazy caching.
}
