% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientforest_ext.R
\name{gf_clust_opt}
\alias{gf_clust_opt}
\title{Collection of GF models F-ratio}
\usage{
gf_clust_opt(gf_list, combined_args = list(standardize = c("before",
  "after")[2]), env_grid, spatial_vars, gf_predict_args, extrap_pow,
  gf_sites, k_range, cluster_parallel = FALSE, clara_args,
  mvpart_args = list(xv = c("1se", "min")[1], xval = 10, xvmult = 10,
  xvse = 1, plot.add = FALSE, text.add = FALSE, pretty = FALSE))
}
\arguments{
\item{gf_list}{list of gradient Forest models. Each element must be named}

\item{env_grid}{data.frame of environmental conditions}

\item{spatial_vars}{character vector specifying columns in env_data that give spatial location rather than predictors}

\item{gf_predict_args}{list of addidional arguments to predict.combinedGradientForest}

\item{extrap_pow}{compression for Gradient Forest extrapolation. 0 for capping, 1 for linear, (0,1) for compression}

\item{gf_sites}{list, one entry per GF object, mapping from each sample site to a row in env_data. Each element must be named to match gf_list}

\item{k_range}{integer vector of cluster k values to fit}

\item{clara_args}{list of arguments to pass to cluster::clara}

\item{mvpart_args}{list of arguments to pass to mvpart::mvpart}

\item{combine_args}{list of additional arguments to combinedGradientForest}
}
\value{
list of gf_combined object, clusterings, long data.frame with all results:
for each gf_oject (including mean) and k: fratio, confusion matrix, mvpart (duplicated for each k and sum for mean)
}
\description{
This function is a collection of actions:
1. Get a list of GF objects to estimate k
2. Combine the GF objects
3. transform env using combined GF
4. Cluster all possible values
5. Given that cluster, calculate F-ratio for each individual GF object for each k
6. Show all f-ratios, and create average, possibly weighted
7. Show all confusion matrix scores, possibly weighted ()
7. Also show mvpart suggestions for each individual GF object
}
